% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPval.R
\name{formatPval}
\alias{formatPval}
\title{Pvalue formatting}
\usage{
formatPval(pval)
}
\arguments{
\item{pval}{pvalue to be formatted}
}
\value{
formatted p-value
}
\description{
formatPval formats p-values. If pvalue is less than 0.0001 it is shown
as the string "<0.0001". Else it is rounded to 4 decimals.
}

