% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqTable.R
\name{freqTable}
\alias{freqTable}
\title{Univariate and bivariate frequency tables and barplots}
\usage{
freqTable(x, by = NULL, xlabel = NULL, bylabel = NULL, plot = TRUE,
  horizontal = FALSE, printFreq = TRUE, density = TRUE,
  showTable = TRUE, pctBycol = TRUE, title = "")
}
\arguments{
\item{x}{variable to tabulate, which can be interpreted as factor}

\item{by}{variable for crosstabulation, also interpreted as factor}

\item{xlabel}{Label of the variable x, to be used in figures and tables.}

\item{bylabel}{Label of the variable by, to be used in figures and tables.}

\item{plot}{Logical. If TRUE a barplot is printed}

\item{horizontal}{Logical. If TRUE bars are printed horizontally}

\item{printFreq}{Logical. If TRUE frequency values are printed in the barplot}

\item{density}{Logical. if TRUE barplot is printed with relative frequencies}

\item{showTable}{Logical. If TRUE frequency table is printed via pander. If FALSE
frequency table is returned invisibly.}

\item{pctBycol}{Logical. If TRUE relative frequencies in crosstabs are computed
by columns else by rows}
}
\value{
an univariate frequency table with absolute and relative frequencies, and
optionally a barplot
}
\description{
freqTable builds an univariate or bivariate frequency table with
counts and percentages at each combination of factor levels factor level and
optionally shows a barplot.
}
\examples{
symptom=sample(LETTERS[1:4],120,replace=TRUE)
disease=sample(c("Yes","No"),120,replace=TRUE)
freqTable(symptom,by=disease,horizontal=FALSE,density=TRUE,printFreq=TRUE,showTable=TRUE)
freqTable(symptom,horizontal=FALSE,density=TRUE,printFreq=TRUE,showTable=FALSE)

}

