% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary2.R
\name{summary2}
\alias{summary2}
\title{Univariate summary of a variable for each level of a factor variable.}
\usage{
summary2(x, by, xlabel = NULL, bylabel = NULL, plot = TRUE,
  ptiles = c(0.25, 0.75), alphaNorm = 0.05, fullreport = FALSE,
  boxplot = TRUE, histogram = TRUE, rug = TRUE, faceted = TRUE,
  densityCurve = TRUE, normalCurve = FALSE, addMeanLine = TRUE,
  digits = 2)
}
\arguments{
\item{x}{variable to be summarized.}

\item{by}{variable for identifying groups, interpreted as factor}

\item{xlabel}{Label of the variable x, to be used in figures and tables.}

\item{bylabel}{Label of the variable by, to be used in figures and tables.}

\item{plot}{Logical. If TRUE a barplot is printed}

\item{ptiles}{ptiles to print with the median when variable is not normal.}

\item{alphaNorm}{Significance level for testing normality. 0.05 by default.}

\item{boxplot}{Should a boxplot be added under the histograms?}

\item{rug}{must a rug be drawn at the base of the histograms?}

\item{densityCurve}{Logical. Must a density curve be superposed to the histograms?}

\item{normalCurve}{Logical. Must a normal curve be superposed to the histograms?}

\item{fullreport.}{If FALSE only mean and sd (when data are normally distributed) or
median and quartiles (when data are not normal) are shown. In other case a full report is printed,
including also number of missing values, skewness, kurtosis, min, max and the p-value of the Shapiro
test for normality}

\item{histColor}{colors for plotting the histograms.}
}
\value{

}
\description{
summary2 makes a simple description of a variable in two or more groups.
It first tests if the variable has a normal distribution by a Shapiro-Wilk test. In the
normal case the variable is summarized in mean and sd in each group. In other case it is
summarized in median and percentiles (by default 25 and 75 percentiles). A comparison
betweed groups is made by a t.test in the normal case and by a wilcoxon test in other case.
Optionally histograms and boxplots are printed.
}

