% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Univariate frequency table and barplot}
\usage{
table1(x, xlabel = NULL, plot = TRUE, horizontal = FALSE,
  printFreq = TRUE, density = TRUE, showTable = TRUE)
}
\arguments{
\item{x}{Variable to tabulate, which can be interpreted as factor}

\item{xlabel}{Label of the variable x, to be used in figures and tables.}

\item{plot}{Logical. If TRUE a barplot is printed}

\item{horizontal}{Logical. If TRUE bars are printed horizontally}

\item{printFreq}{Logical. If TRUE frequency values are printed in the barplot}

\item{density}{Logical. if TRUE barplot is printed with relative frequencies}

\item{showTable}{Logical. If TRUE frequency table is printed via pander. If FALSE
frequency table is returned invisibly.}
}
\value{
an univariate frequency table with absolute and relative frequencies, and
optionally a barplot
}
\description{
table1 builds a univariate frequency table with counts and percentages
at each factor level, and optionally shows a barplot.
}
\examples{
letters=sample(LETTERS[1:4],120,replace=TRUE)
table1(letters,xlabel="letters in a text",horizontal=TRUE,density=TRUE,printFreq=FALSE, showTable=TRUE)

}

