% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Univariate summary of a variable for each level of a factor variable (if any is specified)}
\usage{
describe(x, by = NULL, xlabel = NULL, bylabel = NULL, plot = TRUE,
  report = "auto", showDescriptives = TRUE)
}
\arguments{
\item{x}{variable to be summarized.}

\item{by}{variable for identifying groups, interpreted as factor}

\item{xlabel}{Label of the variable x, to be used in figures and tables.}

\item{bylabel}{Label of the variable by, to be used in figures and tables.}

\item{plot}{Logical. If TRUE a plot is printed}

\item{report.}{If "auto" a Shapiro test is made; only mean and sd (if variable can be assumed to be}
}
\value{

}
\description{
describe makes a simple description of a variable 'x' in one or more groups defined
by the factor 'by'. If 'x' is a factor, frequency tables and (optionally) barplots are shown. If 'x'
is a numeric variable, the variable is summarized in mean and sd in each group if 'x' is normal.
In other case it is summarized in median and percentiles (by default 25 and 75 percentiles). A comparison
betweed two groups is made by a t.test in the normal case and by a wilcoxon test in other case.
When there are more than two groups, anova F-test and Kruskal-Wallis test are used.
Optionally histograms and boxplots are printed.
}

