% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary1.R
\name{summary1}
\alias{summary1}
\title{Univariate summary}
\usage{
summary1(x, xlabel = NULL, plot = TRUE, ptiles = c(0.25, 0.75),
  alphaNorm = 0.05, normalCurve = FALSE)
}
\arguments{
\item{x}{variable to be summarized.}

\item{xlabel}{Label of the variable x, to be used in figures and tables.}

\item{plot}{Logical. If TRUE a barplot is printed}

\item{ptiles}{ptiles to print with the median when variable is not normal.}

\item{alphaNorm}{Significance level for testing normality. 0.05 by default.}

\item{normalCurve}{Logical. Must a normal curve be superposed to the histogram?.}
}
\value{
a brief summary of a continuous variable and optionally an histogram.
}
\description{
summary1 makes a simple description of a variable. It first tests if
variable has a normal distribution by a Shapiro-Wilk test. In the normal case the
variable is summarized in mean and sd. In other case it is summarized in median
and percentiles (by default 25 and 75 percentiles). Optionally an histogram is
printed.
}

