% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary1.R
\name{summary1}
\alias{summary1}
\title{Univariate summary}
\usage{
summary1(x, xlabel = NULL, plot = TRUE, ptiles = c(0.25, 0.75),
  alphaNorm = 0.05, fullreport = FALSE, normalCurve = FALSE,
  densityCurve = TRUE, rug = TRUE, boxplot = TRUE, addMeanLine = TRUE,
  histColor = "#00A4FF")
}
\arguments{
\item{x}{variable to be summarized.}

\item{xlabel}{Label of the variable x, to be used in figures and tables.}

\item{plot}{Logical. If TRUE a barplot is printed}

\item{ptiles}{ptiles to print with the median when variable is not normal.}

\item{alphaNorm}{Significance level for testing normality. 0.05 by default.}

\item{normalCurve}{Logical. Must a normal curve be superposed to the histogram?}

\item{densityCurve}{Logical. Must a density curve be superposed to the histogram?}

\item{rug}{must a rug be drawn at the base of the histogram?}

\item{boxplot}{Should a boxplot be added under the histogram?}

\item{addMeanLine}{Should a vertical line be drawn at the position of the mean?}

\item{histColor}{color for plotting the histogram.}

\item{report.}{If "simple" only mean and sd (when data are normally distributed) or
median and quartiles (when data are not normal) are shown. In other case a full report is printed,
including also number of missing values, skewness, kurtosis, min, max and the p-value of the Shapiro
test for normality}
}
\value{
a brief summary of a continuous variable and optionally an histogram.
}
\description{
summary1 makes a simple description of a variable. It first tests if
variable has a normal distribution by a Shapiro-Wilk test. In the normal case the
variable is summarized in mean and sd. In other case it is summarized in median
and percentiles (by default 25 and 75 percentiles). Optionally an histogram is
printed.
}

